pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "ptpd-calibration"
version = "1.0.0"
description = "AI-powered calibration system for platinum/palladium printing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Ian Shank", email = "ianshank@users.noreply.github.com" }
]
keywords = [
    "platinum",
    "palladium",
    "printing",
    "calibration",
    "photography",
    "alternative-process",
    "ai",
    "machine-learning"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pillow>=10.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.3.0",
]
llm = [
    "anthropic>=0.39.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
]
api = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
]
ui = [
    "gradio>=4.0.0",
]
all = [
    "ptpd-calibration[ml,llm,api,ui]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
ptpd = "ptpd_calibration.cli:main"
ptpd-server = "ptpd_calibration.api.server:main"

[project.urls]
Homepage = "https://github.com/ianshank/Platinum-Palladium-AI-Printing-Tool"
Documentation = "https://github.com/ianshank/Platinum-Palladium-AI-Printing-Tool#readme"
Repository = "https://github.com/ianshank/Platinum-Palladium-AI-Printing-Tool"
Issues = "https://github.com/ianshank/Platinum-Palladium-AI-Printing-Tool/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ptpd_calibration"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src/ptpd_calibration --cov-report=term-missing"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["ptpd_calibration"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
exclude = ["tests"]
